Split TitleDetailPage into per-type and per-section components
Closes #494

The 1033-line TitleDetailPage.tsx is now a slim 65-line router that
fetches details and dispatches to MovieDetail or ShowDetail. Each per-type
container composes small, focused section components from the new
`components/title-detail/` module so individual sections can later be
wrapped in error boundaries and tested in isolation.

Visual parity is preserved: the markup, classes, and prop wiring inside
each section are unchanged.

New structure:
  pages/TitleDetailPage.tsx                       # router only
  pages/title/MovieDetail.tsx                     # movie container
  pages/title/ShowDetail.tsx                      # show container
  components/title-detail/Hero.tsx                # re-exports MovieHero/ShowHero
  components/title-detail/MovieHero.tsx
  components/title-detail/ShowHero.tsx
  components/title-detail/Cast.tsx
  components/title-detail/Crew.tsx
  components/title-detail/ReleaseDates.tsx
  components/title-detail/ProvidersSection.tsx    # WhereToWatch + chips
  components/title-detail/RatingsSection.tsx      # your-rating + share/recommend
  components/title-detail/NetworkList.tsx
  components/title-detail/RatingBadge.tsx
  components/title-detail/Section.tsx             # shared section heading wrapper
  components/title-detail/utils.ts                # date/runtime/currency helpers
  components/title-detail/ProvidersSection.test.tsx
