+++ frontend/src/pages/TitleDetailPage.tsx
@@ -1,240 +1,10 @@
-import { useState, useEffect } from "react";
-import { useParams, Link } from "react-router";
-import { useTranslation } from "react-i18next";
-import ScrollableRow from "../components/ScrollableRow";
+import { useEffect, useState } from "react";
+import { useParams } from "react-router";
 import * as api from "../api";
-import type {
-  Title,
-  MovieDetailsResponse,
-  ShowDetailsResponse,
-  CastMember,
-  CrewMember,
-  WatchProviderCountry,
-  ReleaseDatesResult,
-  SeasonSummary,
-} from "../types";
-import type { WatchHistoryEntry } from "../types";
-import TrackButton from "../components/TrackButton";
-import RatingButtons from "../components/RatingButtons";
-import { useIsMobile } from "../hooks/useIsMobile";
-import PersonCard from "../components/PersonCard";
+import type { MovieDetailsResponse, ShowDetailsResponse } from "../types";
 import { DetailPageSkeleton } from "../components/SkeletonComponents";
-import ExternalLinks from "../components/ExternalLinks";
-import { PLEX_PROVIDER_ID, getPlexPlatform, plexDeepLink } from "../components/WatchButton";
-import WatchButtonGroup from "../components/WatchButtonGroup";
-import VisibilityButton from "../components/VisibilityButton";
-import ShareButton from "../components/ShareButton";
-import RecommendButton from "../components/RecommendButton";
-import BackdateWatchedButton from "../components/BackdateWatchedButton";
-import { getProviderColor } from "../data/providerColors";
-import { Kicker, Chip } from "../components/design";
-
-const TMDB_IMG = "https://image.tmdb.org/t/p";
-
-const RELEASE_TYPE_LABELS: Record<number, string> = {
-  1: "Premiere",
-  2: "Theatrical (Limited)",
-  3: "Theatrical",
-  4: "Digital",
-  5: "Physical",
-  6: "TV",
-};
-
-function formatDate(dateStr: string | null | undefined): string {
-  if (!dateStr) return "—";
-  const d = new Date(dateStr.includes("T") ? dateStr : dateStr + "T00:00:00");
-  if (isNaN(d.getTime())) return "—";
-  return d.toLocaleDateString("en-US", { year: "numeric", month: "short", day: "numeric" });
-}
-
-function formatRuntime(minutes: number | null | undefined): string {
-  if (!minutes) return "";
-  const h = Math.floor(minutes / 60);
-  const m = minutes % 60;
-  return h > 0 ? `${h}h ${m}m` : `${m}m`;
-}
-
-function formatCurrency(value: number): string {
-  if (!value) return "—";
-  return new Intl.NumberFormat("en-US", { style: "currency", currency: "USD", maximumFractionDigits: 0 }).format(value);
-}
-
-function isToday(dateStr: string | null | undefined): boolean {
-  if (!dateStr) return false;
-  const d = new Date(dateStr.includes("T") ? dateStr : dateStr + "T00:00:00");
-  const now = new Date();
-  return d.getFullYear() === now.getFullYear() && d.getMonth() === now.getMonth() && d.getDate() === now.getDate();
-}
-
-// ─── Sub-components ──────────────────────────────────────────────────────────
-
-function RatingBadge({ label, score, max = 10 }: { label: string; score: number | null; max?: number }) {
-  if (score === null || score === undefined) return null;
-  return (
-    <div className="flex flex-col items-center gap-1">
-      <span className="text-[11px] font-medium text-zinc-400 uppercase tracking-[0.1em]">{label}</span>
-      <span className="text-[22px] font-bold text-white leading-none">
-        {score.toFixed(1)}<span className="text-sm text-zinc-500">/{max}</span>
-      </span>
-    </div>
-  );
-}
-
-function Section({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <section className="space-y-4">
-      <h2 className="text-[18px] font-semibold text-white tracking-tight leading-tight">{title}</h2>
-      {children}
-    </section>
-  );
-}
-
-const NETWORK_DISPLAY_LIMIT = 5;
-
-function NetworkList({ networks }: { networks: { id: number; name: string; logo_path: string | null }[] }) {
-  const [expanded, setExpanded] = useState(false);
-  const hasMore = networks.length > NETWORK_DISPLAY_LIMIT;
-  const visible = expanded ? networks : networks.slice(0, NETWORK_DISPLAY_LIMIT);
-  return (
-    <div className="flex flex-wrap items-center gap-3">
-      {visible.map(n => (
-        <div key={n.id} className="flex items-center gap-1.5">
-          {n.logo_path && (
-            <img src={`${TMDB_IMG}/w92${n.logo_path}`} alt={n.name} className="h-5 object-contain brightness-0 invert opacity-70" />
-          )}
-          <span className="text-sm text-zinc-400">{n.name}</span>
-        </div>
-      ))}
-      {hasMore && (
-        <button
-          onClick={() => setExpanded(!expanded)}
-          className="text-sm text-amber-400 hover:text-amber-300 transition-colors"
-        >
-          {expanded ? "Show less" : `+${networks.length - NETWORK_DISPLAY_LIMIT} more`}
-        </button>
-      )}
-    </div>
-  );
-}
-
-function ProviderRow({ label, children, isEmpty }: { label: string; children: React.ReactNode; isEmpty: boolean }) {
-  return (
-    <div className="flex items-start gap-3.5">
-      <div className="w-[70px] shrink-0 pt-1.5 text-sm text-zinc-400">{label}</div>
-      <div className="flex flex-wrap gap-2 items-center min-h-8">
-        {isEmpty ? (
-          <span className="text-xs text-zinc-600 italic">— not available</span>
-        ) : (
-          children
-        )}
-      </div>
-    </div>
-  );
-}
-
-function OfferChip({ offer }: { offer: Title["offers"][number] }) {
-  const color = getProviderColor(offer.provider_id);
-  const isPlex = offer.provider_id === PLEX_PROVIDER_ID;
-  const platform = isPlex ? getPlexPlatform() : "desktop";
-  const useMobileDeepLink = platform === "ios" || platform === "android";
-  const url = useMobileDeepLink ? plexDeepLink(offer.url, platform) : offer.url;
-  return (
-    <a
-      href={url}
-      target={useMobileDeepLink ? undefined : "_blank"}
-      rel="noopener noreferrer"
-      className="inline-flex items-center gap-2 rounded-lg px-2.5 py-1.5 hover:opacity-90 transition-opacity"
-      style={{ backgroundColor: `${color.bg}20`, borderLeft: `3px solid ${color.bg}` }}
-      title={offer.provider_name}
-    >
-      <img
-        src={offer.provider_icon_url}
-        alt={offer.provider_name}
-        className="w-6 h-6 rounded"
-        loading="lazy"
-      />
-      <span className="text-sm text-zinc-300">{offer.provider_name}</span>
-    </a>
-  );
-}
-
-function TmdbProviderChip({
-  provider,
-  watchLink,
-}: {
-  provider: { logo_path: string; provider_name: string; provider_id: number };
-  watchLink?: string;
-}) {
-  const color = getProviderColor(provider.provider_id);
-  const chip = (
-    <div
-      className="inline-flex items-center gap-2 rounded-lg px-2.5 py-1.5"
-      style={{ backgroundColor: `${color.bg}20`, borderLeft: `3px solid ${color.bg}` }}
-    >
-      <img src={`${TMDB_IMG}/w45${provider.logo_path}`} alt={provider.provider_name} className="w-6 h-6 rounded" />
-      <span className="text-sm text-zinc-300">{provider.provider_name}</span>
-    </div>
-  );
-  if (watchLink) {
-    return (
-      <a href={watchLink} target="_blank" rel="noopener noreferrer" className="hover:opacity-90 transition-opacity">
-        {chip}
-      </a>
-    );
-  }
-  return chip;
-}
-
-function WhereToWatch({
-  offers,
-  watchProviders,
-  watchLink,
-}: {
-  offers: Title["offers"];
-  watchProviders: WatchProviderCountry | undefined;
-  watchLink: string | undefined;
-}) {
-  const offerGroups = groupOffersByType(offers);
-  const hasOffers = offerGroups.length > 0;
-  const hasProviders = !!watchProviders;
-  if (!hasOffers && !hasProviders) return null;
-
-  const providerMap: Record<"FLATRATE" | "FREE" | "ADS" | "RENT" | "BUY", "flatrate" | "free" | "ads" | "rent" | "buy"> = {
-    FLATRATE: "flatrate",
-    FREE: "free",
-    ADS: "ads",
-    RENT: "rent",
-    BUY: "buy",
-  };
-
-  return (
-    <Section title="Where to Watch">
-      <div className="flex flex-col gap-3 max-w-4xl">
-        {MONETIZATION_ORDER.map(({ type, label }) => {
-          const groupOffers = hasOffers
-            ? (offerGroups.find((g) => g.type === type)?.offers ?? [])
-            : [];
-          const tmdbKey = providerMap[type];
-          const tmdbProviders =
-            !hasOffers && hasProviders && watchProviders ? (watchProviders[tmdbKey] ?? []) : [];
-          const isEmpty = groupOffers.length === 0 && tmdbProviders.length === 0;
-          return (
-            <ProviderRow key={type} label={label} isEmpty={isEmpty}>
-              {groupOffers.map((offer) => (
-                <OfferChip key={offer.id} offer={offer} />
-              ))}
-              {tmdbProviders.map((p) => (
-                <TmdbProviderChip key={p.provider_id} provider={p} watchLink={watchLink} />
-              ))}
-            </ProviderRow>
-          );
-        })}
-      </div>
-    </Section>
-  );
-}
-
-// ─── Main Component ──────────────────────────────────────────────────────────
+import MovieDetail from "./title/MovieDetail";
+import ShowDetail from "./title/ShowDetail";
 
 export default function TitleDetailPage() {
   const { id } = useParams<{ id: string }>();
@@ -271,7 +41,9 @@ export default function TitleDetailPage() {
     }
 
     load();
-    return () => { cancelled = true; };
+    return () => {
+      cancelled = true;
+    };
   }, [id]);
 
   if (loading) {
@@ -290,744 +62,4 @@ export default function TitleDetailPage() {
   if (showData) return <ShowDetail data={showData} />;
 
   return <div className="text-zinc-400 text-center py-20">Title not found</div>;
-}
-
-// ─── Movie Detail ────────────────────────────────────────────────────────────
-
-function MovieDetail({ data }: { data: MovieDetailsResponse }) {
-  const { t } = useTranslation();
-  const { title, tmdb, country } = data;
-  const [watched, setWatched] = useState(title.is_watched ?? false);
-  const [playCount, setPlayCount] = useState(0);
-  const [watchHistory, setWatchHistory] = useState<WatchHistoryEntry[]>([]);
-  const [historyOpen, setHistoryOpen] = useState(false);
-
-  useEffect(() => {
-    api.getWatchHistory(title.id).then(({ history, playCount: cnt }) => {
-      setWatchHistory(history);
-      setPlayCount(cnt);
-    }).catch(() => {});
-  }, [title.id]);
-
-  async function toggleWatched() {
-    const prev = watched;
-    setWatched(!prev);
-    try {
-      if (prev) {
-        await api.unwatchMovie(title.id);
-      } else {
-        await api.watchMovie(title.id);
-        // Refresh history after marking watched
-        const { history, playCount: cnt } = await api.getWatchHistory(title.id);
-        setWatchHistory(history);
-        setPlayCount(cnt);
-      }
-    } catch {
-      setWatched(prev);
-    }
-  }
-  const overview = tmdb?.overview || title.short_description;
-  const genres = tmdb?.genres?.map(g => g.name) || title.genres;
-  const certification = title.age_certification;
-  const backdropUrl = tmdb?.backdrop_path ? `${TMDB_IMG}/w1280${tmdb.backdrop_path}` : null;
-  const posterUrl = tmdb?.poster_path ? `${TMDB_IMG}/w500${tmdb.poster_path}` : title.poster_url;
-
-  // Release dates for the user's country
-  const countryReleaseDates = tmdb?.release_dates?.results?.find(
-    (r: ReleaseDatesResult) => r.iso_3166_1 === country
-  );
-  // Also check US as fallback
-  const usReleaseDates = tmdb?.release_dates?.results?.find(
-    (r: ReleaseDatesResult) => r.iso_3166_1 === "US"
-  );
-  const releaseDates = countryReleaseDates || usReleaseDates;
-
-  // Watch providers for the user's country
-  const watchProviders = tmdb?.["watch/providers"]?.results?.[country] as WatchProviderCountry | undefined;
-
-  // Key crew
-  const directors = tmdb?.credits?.crew?.filter((c: CrewMember) => c.job === "Director") || [];
-  const writers = tmdb?.credits?.crew?.filter((c: CrewMember) => c.department === "Writing").slice(0, 5) || [];
-  const cast = tmdb?.credits?.cast?.slice(0, 20) || [];
-
-  return (
-    <div className="space-y-8 pb-12">
-      {/* Hero */}
-      <div className="relative -mx-4 -mt-6 px-4 pt-6 pb-8 sm:px-8 sm:pt-10 sm:pb-10 lg:px-16 lg:pt-14 lg:pb-12 dark-section" style={backdropUrl ? {
-        backgroundImage: `linear-gradient(to bottom, rgba(3,7,18,0.6), rgba(3,7,18,1)), url(${backdropUrl})`,
-        backgroundSize: "cover",
-        backgroundPosition: "center top",
-      } : undefined}>
-        <div className="flex flex-col sm:flex-row gap-6 sm:gap-8 lg:gap-10 items-end">
-          {/* Poster */}
-          <div className="w-48 sm:w-56 lg:w-60 shrink-0 mx-auto sm:mx-0">
-            {posterUrl ? (
-              <img src={posterUrl} alt={title.title} className="w-full rounded-xl shadow-[0_24px_70px_rgba(0,0,0,0.7)] border border-white/[0.08]" />
-            ) : (
-              <div className="aspect-[2/3] bg-zinc-800 rounded-xl flex items-center justify-center text-zinc-600 border border-white/[0.08]">No poster</div>
-            )}
-          </div>
-
-          {/* Title info */}
-          <div className="flex-1 space-y-3 max-w-[820px]">
-            <div>
-              {tmdb?.tagline && (
-                <p className="text-sm text-amber-400 italic mb-1">{tmdb.tagline}</p>
-              )}
-              <Kicker className="mb-2">
-                Movie{title.release_year ? ` · ${title.release_year}` : ""}{title.offers[0]?.provider_name ? ` · ${title.offers[0].provider_name}` : ""}
-              </Kicker>
-              <h1 className="text-[30px] sm:text-[56px] lg:text-[64px] leading-none tracking-[-0.035em] font-extrabold text-white">{tmdb?.title || title.title}</h1>
-              {(() => {
-                const displayTitle = tmdb?.title || title.title;
-                const originalTitle = tmdb?.original_title || title.original_title;
-                return originalTitle && originalTitle !== displayTitle ? (
-                  <p className="text-sm text-zinc-400 mt-1">{originalTitle}</p>
-                ) : null;
-              })()}
-            </div>
-
-            <div className="flex flex-wrap items-center gap-2 text-sm text-zinc-400">
-              {title.release_year && <span>{title.release_year}</span>}
-              {(title.runtime_minutes || tmdb?.runtime) && (
-                <>
-                  <span className="text-zinc-600">·</span>
-                  <span>{formatRuntime(tmdb?.runtime || title.runtime_minutes)}</span>
-                </>
-              )}
-              {certification && (
-                <>
-                  <span className="text-zinc-600">·</span>
-                  <span className="border border-white/[0.10] px-1.5 py-0.5 rounded text-xs">{certification}</span>
-                </>
-              )}
-              {tmdb?.status && (
-                <>
-                  <span className="text-zinc-600">·</span>
-                  <span>{tmdb.status}</span>
-                </>
-              )}
-            </div>
-
-            {(genres.length > 0 || title.imdb_score) && (
-              <div className="flex flex-wrap gap-2">
-                {genres.map((g) => (
-                  <Chip key={g} variant="default">{g}</Chip>
-                ))}
-                {title.imdb_score && (
-                  <Chip variant="amber">★ {title.imdb_score.toFixed(1)}</Chip>
-                )}
-              </div>
-            )}
-
-            {/* Ratings */}
-            <div className="flex items-center gap-6 pt-2">
-              <RatingBadge label="IMDb" score={title.imdb_score} />
-              <RatingBadge label="TMDB" score={tmdb?.vote_average ?? title.tmdb_score} />
-            </div>
-
-            <div className="pt-2 flex flex-wrap items-center gap-2">
-              <TrackButton titleId={title.id} isTracked={title.is_tracked} titleData={title} />
-              <VisibilityButton titleId={title.id} isPublic={title.is_public ?? true} isTracked={title.is_tracked} />
-              <button
-                onClick={toggleWatched}
-                aria-pressed={watched}
-                className={`min-h-8 inline-flex items-center justify-center px-3 py-1.5 rounded-md text-xs font-medium transition-colors cursor-pointer ${
-                  watched
-                    ? "bg-emerald-500 text-white hover:bg-red-500"
-                    : "bg-zinc-800 text-zinc-400 hover:bg-emerald-500 hover:text-white"
-                }`}
-              >
-                {watched ? t("episodes.markAsUnwatched") : t("episodes.markAsWatched")}
-              </button>
-              {playCount > 0 && (
-                <button
-                  onClick={() => setHistoryOpen((o) => !o)}
-                  className="min-h-8 inline-flex items-center gap-1 px-3 py-1.5 rounded-md text-xs font-medium bg-zinc-800 text-zinc-400 hover:bg-zinc-700 transition-colors cursor-pointer"
-                  title="Watch history"
-                >
-                  <span>Watched {playCount}x</span>
-                  <span className="text-zinc-600">{historyOpen ? "▲" : "▼"}</span>
-                </button>
-              )}
-              <WatchButtonGroup offers={title.offers} variant="inline" maxVisible={3} />
-            </div>
-            {historyOpen && watchHistory.length > 0 && (
-              <div className="mt-3 rounded-lg bg-zinc-900/60 border border-white/[0.06] overflow-hidden">
-                <div className="px-3 py-2 text-xs font-medium text-zinc-400 border-b border-white/[0.06]">Watch History</div>
-                <ul>
-                  {watchHistory.map((entry, i) => (
-                    <li key={entry.id} className={`px-3 py-2 text-xs text-zinc-300 flex items-center gap-2 ${i < watchHistory.length - 1 ? "border-b border-zinc-800/50" : ""}`}>
-                      <span className="text-zinc-500 shrink-0">{new Date(entry.watchedAt).toLocaleDateString("en-US", { year: "numeric", month: "short", day: "numeric" })}</span>
-                      {entry.note && <span className="text-zinc-400 italic">{entry.note}</span>}
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-
-      {/* Metadata strip */}
-      <div className="dark-section -mx-4 px-6 sm:px-12 py-5 flex flex-wrap gap-x-10 gap-y-3 border-b border-white/[0.06]">
-        {[
-          { label: 'TYPE', value: 'Movie' },
-          title.runtime_minutes
-            ? { label: 'RUNTIME', value: `${title.runtime_minutes} min` }
-            : null,
-          tmdb?.status
-            ? { label: 'STATUS', value: tmdb.status }
-            : null,
-          title.offers[0]?.provider_name
-            ? { label: 'NETWORK', value: title.offers[0].provider_name }
-            : null,
-          title.imdb_score
-            ? { label: 'IMDB', value: `★ ${title.imdb_score.toFixed(1)}` }
-            : null,
-        ].filter(Boolean).map(cell => (
-          <div key={cell!.label}>
-            <div className="font-mono text-[10px] uppercase tracking-[0.12em] text-zinc-500 mb-1">{cell!.label}</div>
-            <div className="font-mono text-[13px] font-semibold text-zinc-100">{cell!.value}</div>
-          </div>
-        ))}
-      </div>
-
-      {/* Overview */}
-      {overview && (
-        <Section title="Overview">
-          <p className="text-zinc-300 leading-relaxed">{overview}</p>
-        </Section>
-      )}
-
-      {/* Rating & Social */}
-      <div className="flex flex-wrap items-start gap-6">
-        <div className="space-y-2">
-          <Kicker color="zinc" className="mb-0">Your rating</Kicker>
-          <RatingButtons titleId={title.id} />
-        </div>
-        <div className="flex-1 min-w-0" />
-        <div className="flex items-center gap-2 pt-6">
-          <ShareButton title={tmdb?.title || title.title} />
-          <RecommendButton titleId={title.id} />
-        </div>
-      </div>
-
-      {/* Cast & Crew */}
-      {(directors.length > 0 || writers.length > 0) && (
-        <Section title="Crew">
-          <div className="flex flex-wrap gap-x-6 gap-y-1 text-sm">
-            {directors.length > 0 && (
-              <div>
-                <span className="text-zinc-400">Director: </span>
-                <span className="text-white">{directors.map(d => d.name).join(", ")}</span>
-              </div>
-            )}
-            {writers.length > 0 && (
-              <div>
-                <span className="text-zinc-400">Writers: </span>
-                <span className="text-white">{writers.map(w => w.name).join(", ")}</span>
-              </div>
-            )}
-          </div>
-        </Section>
-      )}
-
-      {cast.length > 0 && (
-        <Section title="Cast">
-          <ScrollableRow className="gap-4 pb-2">
-            {cast.map((c: CastMember) => (
-              <PersonCard key={c.id} id={c.id} name={c.name} role={c.character} profilePath={c.profile_path} />
-            ))}
-          </ScrollableRow>
-        </Section>
-      )}
-
-      {/* Release Dates */}
-      {releaseDates && releaseDates.release_dates.length > 0 && (
-        <Section title={`Release Dates (${releaseDates.iso_3166_1})`}>
-          <div className="bg-zinc-900 rounded-xl border border-white/[0.06] overflow-hidden">
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="border-b border-white/[0.06]">
-                  <th className="text-left py-2 px-4 text-zinc-400 font-medium">Type</th>
-                  <th className="text-left py-2 px-4 text-zinc-400 font-medium">Date</th>
-                  <th className="text-left py-2 px-4 text-zinc-400 font-medium">Certification</th>
-                  <th className="text-left py-2 px-4 text-zinc-400 font-medium">Note</th>
-                </tr>
-              </thead>
-              <tbody>
-                {releaseDates.release_dates.map((rd) => (
-                  <tr key={`${rd.release_date}-${rd.type}`} className="border-b border-zinc-800/50 last:border-0">
-                    <td className="py-2 px-4 text-zinc-300">{RELEASE_TYPE_LABELS[rd.type] || `Type ${rd.type}`}</td>
-                    <td className="py-2 px-4 text-zinc-300">{formatDate(rd.release_date)}</td>
-                    <td className="py-2 px-4">
-                      {rd.certification && (
-                        <span className="border border-white/[0.10] px-1.5 py-0.5 rounded text-xs text-zinc-300">{rd.certification}</span>
-                      )}
-                    </td>
-                    <td className="py-2 px-4 text-zinc-500 text-xs">{rd.note}</td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </Section>
-      )}
-
-      {/* Streaming Availability */}
-      <WhereToWatch offers={title.offers} watchProviders={watchProviders} watchLink={watchProviders?.link} />
-
-      {/* External Links */}
-      {tmdb && (
-        <Section title="Links">
-          <ExternalLinks
-            externalIds={tmdb.external_ids ?? { imdb_id: tmdb.imdb_id }}
-            tmdbId={tmdb.id}
-            type="movie"
-          />
-        </Section>
-      )}
-
-      {/* Additional Info */}
-      {tmdb && (
-        <Section title="Details">
-          <div className="grid grid-cols-2 sm:grid-cols-3 gap-4 text-sm">
-            {tmdb.original_language && (
-              <div>
-                <span className="text-zinc-500 block">Original Language</span>
-                <span className="text-zinc-300">{tmdb.original_language.toUpperCase()}</span>
-              </div>
-            )}
-            {tmdb.production_countries?.length > 0 && (
-              <div>
-                <span className="text-zinc-500 block">Country</span>
-                <span className="text-zinc-300">{tmdb.production_countries.map(c => c.name).join(", ")}</span>
-              </div>
-            )}
-            {tmdb.spoken_languages?.length > 0 && (
-              <div>
-                <span className="text-zinc-500 block">Languages</span>
-                <span className="text-zinc-300">{tmdb.spoken_languages.map(l => l.english_name).join(", ")}</span>
-              </div>
-            )}
-            {tmdb.budget > 0 && (
-              <div>
-                <span className="text-zinc-500 block">Budget</span>
-                <span className="text-zinc-300">{formatCurrency(tmdb.budget)}</span>
-              </div>
-            )}
-            {tmdb.revenue > 0 && (
-              <div>
-                <span className="text-zinc-500 block">Revenue</span>
-                <span className="text-zinc-300">{formatCurrency(tmdb.revenue)}</span>
-              </div>
-            )}
-            {tmdb.production_companies?.length > 0 && (
-              <div className="col-span-2 sm:col-span-3">
-                <span className="text-zinc-500 block">Production</span>
-                <span className="text-zinc-300">{tmdb.production_companies.map(c => c.name).join(", ")}</span>
-              </div>
-            )}
-          </div>
-        </Section>
-      )}
-    </div>
-  );
-}
-
-// ─── Show Detail ─────────────────────────────────────────────────────────────
-
-function ShowDetail({ data }: { data: ShowDetailsResponse }) {
-  const { title, tmdb, country } = data;
-  const isMobile = useIsMobile();
-  const overview = tmdb?.overview || title.short_description;
-  const genres = tmdb?.genres?.map(g => g.name) || title.genres;
-  const certification = getCertification(tmdb?.content_ratings?.results, country) || title.age_certification;
-  const backdropUrl = tmdb?.backdrop_path ? `${TMDB_IMG}/w1280${tmdb.backdrop_path}` : null;
-  const posterUrl = tmdb?.poster_path ? `${TMDB_IMG}/w500${tmdb.poster_path}` : title.poster_url;
-
-  const watchProviders = tmdb?.["watch/providers"]?.results?.[country] as WatchProviderCountry | undefined;
-
-  const creators = tmdb?.created_by || [];
-  const cast = tmdb?.credits?.cast?.slice(0, 20) || [];
-
-  // Filter out specials (season 0) and sort
-  const seasons = (tmdb?.seasons || [])
-    .filter((s: SeasonSummary) => s.season_number > 0)
-    .sort((a: SeasonSummary, b: SeasonSummary) => a.season_number - b.season_number);
-
-  const firstOfferUrl = title.offers[0]?.url ?? null;
-
-  return (
-    <div className="space-y-8 pb-12">
-      {/* Hero */}
-      {isMobile ? (
-        <>
-          {/* Mobile: 460px full-bleed hero with bottom-anchored poster+title */}
-          <div
-            className="relative -mx-4 -mt-6 overflow-hidden"
-            style={{ height: 460 }}
-          >
-            {backdropUrl ? (
-              <img src={backdropUrl} alt="" className="absolute inset-0 w-full h-full object-cover object-top" />
-            ) : (
-              <div className="absolute inset-0 bg-gradient-to-b from-zinc-800 to-zinc-950" />
-            )}
-            <div className="absolute inset-0" style={{ background: "linear-gradient(180deg, rgba(9,9,11,0.3) 0%, rgba(9,9,11,0.1) 35%, #09090b 95%)" }} />
-            {/* Bottom-anchored poster + title row */}
-            <div className="absolute bottom-0 left-0 right-0 flex items-end gap-4 px-4 pb-5">
-              <div className="w-[108px] shrink-0">
-                {posterUrl ? (
-                  <img src={posterUrl} alt={title.title} className="w-full rounded-xl shadow-2xl" />
-                ) : (
-                  <div className="aspect-[2/3] bg-zinc-800 rounded-xl" />
-                )}
-              </div>
-              <div className="flex-1 min-w-0 pb-1">
-                <div className="font-mono text-[10px] uppercase tracking-[0.14em] text-amber-400 font-semibold mb-1">
-                  SHOW · {title.release_year ?? tmdb?.first_air_date?.slice(0, 4) ?? ""}
-                  {title.offers[0]?.provider_name ? ` · ${title.offers[0].provider_name}` : ""}
-                </div>
-                <h1 className="text-[26px] leading-[1.05] font-bold text-white line-clamp-3">{tmdb?.name || title.title}</h1>
-                <div className="flex flex-wrap gap-1.5 mt-2">
-                  {genres.slice(0, 3).map(g => (
-                    <span key={g} className="bg-white/[0.08] border border-white/[0.10] text-zinc-300 text-[11px] px-2 py-0.5 rounded-full">{g}</span>
-                  ))}
-                  {title.imdb_score && (
-                    <span className="bg-amber-400/[0.15] border border-amber-400/[0.3] text-amber-300 text-[11px] px-2 py-0.5 rounded-full">★ {title.imdb_score.toFixed(1)}</span>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-          {/* Mobile CTA row */}
-          <div className="flex gap-2 -mt-2">
-            {firstOfferUrl ? (
-              <a
-                href={firstOfferUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="flex-1 flex items-center justify-center gap-2 bg-amber-400 hover:bg-amber-300 active:bg-amber-500 text-zinc-950 px-4 py-3 rounded-xl text-[14px] font-bold transition-colors"
-              >
-                ▶ Play
-              </a>
-            ) : (
-              <div className="flex-1 flex items-center justify-center gap-2 bg-white/[0.06] border border-white/[0.08] text-zinc-500 px-4 py-3 rounded-xl text-[14px] font-bold cursor-not-allowed">
-                ▶ No stream
-              </div>
-            )}
-            <TrackButton titleId={title.id} isTracked={title.is_tracked} titleData={title} />
-          </div>
-        </>
-      ) : (
-        <div className="relative -mx-4 -mt-6 px-4 pt-6 pb-8 sm:px-8 sm:pt-10 sm:pb-10 lg:px-16 lg:pt-14 lg:pb-12 dark-section" style={backdropUrl ? {
-          backgroundImage: `linear-gradient(to bottom, rgba(3,7,18,0.6), rgba(3,7,18,1)), url(${backdropUrl})`,
-          backgroundSize: "cover",
-          backgroundPosition: "center top",
-        } : undefined}>
-          <div className="flex flex-col sm:flex-row gap-6 sm:gap-8 lg:gap-10 items-end">
-            <div className="w-48 sm:w-56 lg:w-60 shrink-0 mx-auto sm:mx-0">
-              {posterUrl ? (
-                <img src={posterUrl} alt={title.title} className="w-full rounded-xl shadow-[0_24px_70px_rgba(0,0,0,0.7)] border border-white/[0.08]" />
-              ) : (
-                <div className="aspect-[2/3] bg-zinc-800 rounded-xl flex items-center justify-center text-zinc-600 border border-white/[0.08]">No poster</div>
-              )}
-            </div>
-
-            <div className="flex-1 space-y-3 max-w-[820px]">
-              <div>
-                {tmdb?.tagline && (
-                  <p className="text-sm text-amber-400 italic mb-1">{tmdb.tagline}</p>
-                )}
-                <Kicker className="mb-2">
-                  TV Show{title.release_year ? ` · ${title.release_year}` : ""}{title.offers[0]?.provider_name ? ` · ${title.offers[0].provider_name}` : ""}
-                </Kicker>
-                <h1 className="text-[30px] sm:text-[56px] lg:text-[64px] leading-none tracking-[-0.035em] font-extrabold text-white">{tmdb?.name || title.title}</h1>
-                {(() => {
-                  const displayTitle = tmdb?.name || title.title;
-                  const originalTitle = tmdb?.original_name || title.original_title;
-                  return originalTitle && originalTitle !== displayTitle ? (
-                    <p className="text-sm text-zinc-400 mt-1">{originalTitle}</p>
-                  ) : null;
-                })()}
-              </div>
-
-              <div className="flex flex-wrap items-center gap-2 text-sm text-zinc-400">
-                <span className="bg-amber-500 text-black text-xs font-bold px-1.5 py-0.5 rounded">TV</span>
-                {tmdb?.first_air_date && <span>{tmdb.first_air_date.slice(0, 4)}</span>}
-                {tmdb?.last_air_date && tmdb.first_air_date?.slice(0, 4) !== tmdb.last_air_date.slice(0, 4) && (
-                  <span>– {tmdb.last_air_date.slice(0, 4)}</span>
-                )}
-                {tmdb?.episode_run_time?.[0] && (
-                  <>
-                    <span className="text-zinc-600">·</span>
-                    <span>{tmdb.episode_run_time[0]}m/ep</span>
-                  </>
-                )}
-                {certification && (
-                  <>
-                    <span className="text-zinc-600">·</span>
-                    <span className="border border-white/[0.10] px-1.5 py-0.5 rounded text-xs">{certification}</span>
-                  </>
-                )}
-                {tmdb?.status && (
-                  <>
-                    <span className="text-zinc-600">·</span>
-                    <span>{tmdb.status}</span>
-                  </>
-                )}
-                {tmdb && (
-                  <>
-                    <span className="text-zinc-600">·</span>
-                    <span>{tmdb.number_of_seasons} season{tmdb.number_of_seasons !== 1 ? "s" : ""}</span>
-                    <span className="text-zinc-600">·</span>
-                    <span>{tmdb.number_of_episodes} episodes</span>
-                  </>
-                )}
-              </div>
-
-              {(genres.length > 0 || title.imdb_score) && (
-                <div className="flex flex-wrap gap-2">
-                  {genres.map((g) => (
-                    <Chip key={g} variant="default">{g}</Chip>
-                  ))}
-                  {title.imdb_score && (
-                    <Chip variant="amber">★ {title.imdb_score.toFixed(1)}</Chip>
-                  )}
-                </div>
-              )}
-
-              {/* Networks */}
-              {tmdb?.networks && tmdb.networks.length > 0 && (
-                <NetworkList networks={tmdb.networks} />
-              )}
-
-              <div className="flex items-center gap-6 pt-2">
-                <RatingBadge label="IMDb" score={title.imdb_score} />
-                <RatingBadge label="TMDB" score={tmdb?.vote_average ?? title.tmdb_score} />
-              </div>
-
-              <div className="pt-2 flex flex-wrap items-center gap-2">
-                <TrackButton titleId={title.id} isTracked={title.is_tracked} titleData={title} />
-                <VisibilityButton titleId={title.id} isPublic={title.is_public ?? true} isTracked={title.is_tracked} />
-                <WatchButtonGroup offers={title.offers} variant="inline" maxVisible={3} />
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Metadata strip */}
-      <div className="dark-section -mx-4 px-6 sm:px-12 py-5 flex flex-wrap gap-x-10 gap-y-3 border-b border-white/[0.06]">
-        {[
-          { label: 'TYPE', value: 'TV Show' },
-          tmdb?.status
-            ? { label: 'STATUS', value: tmdb.status }
-            : null,
-          tmdb?.number_of_seasons != null
-            ? { label: 'SEASONS', value: String(tmdb.number_of_seasons) }
-            : null,
-          tmdb?.number_of_episodes != null
-            ? { label: 'EPISODES', value: String(tmdb.number_of_episodes) }
-            : null,
-          title.next_episode_air_date
-            ? { label: 'NEXT EP', value: formatDate(title.next_episode_air_date) }
-            : null,
-          title.offers[0]?.provider_name
-            ? { label: 'NETWORK', value: title.offers[0].provider_name }
-            : null,
-          tmdb?.episode_run_time?.[0]
-            ? { label: 'AVG RUNTIME', value: `${tmdb.episode_run_time[0]} min` }
-            : null,
-          title.imdb_score
-            ? { label: 'IMDB', value: `★ ${title.imdb_score.toFixed(1)}` }
-            : null,
-        ].filter(Boolean).map(cell => (
-          <div key={cell!.label}>
-            <div className="font-mono text-[10px] uppercase tracking-[0.12em] text-zinc-500 mb-1">{cell!.label}</div>
-            <div className="font-mono text-[13px] font-semibold text-zinc-100">{cell!.value}</div>
-          </div>
-        ))}
-      </div>
-
-      {/* Overview */}
-      {overview && (
-        <Section title="Overview">
-          <p className="text-zinc-300 leading-relaxed">{overview}</p>
-        </Section>
-      )}
-
-      {/* Rating & Social */}
-      <div className="flex flex-wrap items-start gap-6">
-        <div className="space-y-2">
-          <Kicker color="zinc" className="mb-0">Your rating</Kicker>
-          <RatingButtons titleId={title.id} />
-        </div>
-        <div className="flex-1 min-w-0" />
-        <div className="flex items-center gap-2 pt-6">
-          <ShareButton title={tmdb?.name || title.title} />
-          <RecommendButton titleId={title.id} />
-        </div>
-      </div>
-
-      {/* Creators & Cast */}
-      {creators.length > 0 && (
-        <Section title="Created By">
-          <div className="flex gap-4">
-            {creators.map((c) => (
-              <PersonCard key={c.id} id={c.id} name={c.name} role="Creator" profilePath={c.profile_path} />
-            ))}
-          </div>
-        </Section>
-      )}
-
-      {cast.length > 0 && (
-        <Section title="Cast">
-          <ScrollableRow className="gap-4 pb-2">
-            {cast.map((c: CastMember) => (
-              <PersonCard key={c.id} id={c.id} name={c.name} role={c.character} profilePath={c.profile_path} />
-            ))}
-          </ScrollableRow>
-        </Section>
-      )}
-
-      {/* Seasons */}
-      {seasons.length > 0 && (
-        <section className="space-y-4">
-          <div className="flex items-center justify-between gap-3 flex-wrap">
-            <h2 className="text-[18px] font-semibold text-white tracking-tight leading-tight">Seasons</h2>
-            <BackdateWatchedButton scope="title" titleId={title.id} variant="ghost" />
-          </div>
-          <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 gap-4">
-            {seasons.map((s: SeasonSummary) => {
-              const airingToday = isToday(s.air_date);
-              return (
-                <Link
-                  key={s.season_number}
-                  to={`/title/${title.id}/season/${s.season_number}`}
-                  className={`rounded-xl overflow-hidden border transition-colors group ${
-                    airingToday
-                      ? "bg-amber-400/[0.06] border-amber-400/25 hover:border-amber-400/50"
-                      : "bg-zinc-900 border-white/[0.06] hover:border-amber-500/50"
-                  }`}
-                >
-                  <div className="aspect-[2/3] bg-zinc-800">
-                    {s.poster_path ? (
-                      <img
-                        src={`${TMDB_IMG}/w342${s.poster_path}`}
-                        alt={s.name}
-                        className="w-full h-full object-cover"
-                        loading="lazy"
-                      />
-                    ) : (
-                      <div className="w-full h-full flex items-center justify-center text-zinc-600 text-sm">
-                        Season {s.season_number}
-                      </div>
-                    )}
-                  </div>
-                  <div className="p-3">
-                    <div className="flex items-center gap-1.5">
-                      <h3 className="text-sm font-medium text-white group-hover:text-amber-400 transition-colors truncate">{s.name}</h3>
-                      {airingToday && (
-                        <span className="font-mono text-[10px] text-amber-400 tracking-[0.12em] uppercase shrink-0">AIRING NOW</span>
-                      )}
-                    </div>
-                    <p className="text-xs text-zinc-500 mt-0.5">
-                      {s.episode_count} episode{s.episode_count !== 1 ? "s" : ""}
-                      {s.air_date && ` · ${s.air_date.slice(0, 4)}`}
-                    </p>
-                  </div>
-                </Link>
-              );
-            })}
-          </div>
-        </section>
-      )}
-
-      {/* Streaming Availability */}
-      <WhereToWatch offers={title.offers} watchProviders={watchProviders} watchLink={watchProviders?.link} />
-
-      {/* External Links */}
-      {tmdb && (
-        <Section title="Links">
-          <ExternalLinks
-            externalIds={tmdb.external_ids}
-            tmdbId={tmdb.id}
-            type="tv"
-          />
-        </Section>
-      )}
-
-      {/* Additional Info */}
-      {tmdb && (
-        <Section title="Details">
-          <div className="grid grid-cols-2 sm:grid-cols-3 gap-4 text-sm">
-            {tmdb.type && (
-              <div>
-                <span className="text-zinc-500 block">Type</span>
-                <span className="text-zinc-300">{tmdb.type}</span>
-              </div>
-            )}
-            {tmdb.original_language && (
-              <div>
-                <span className="text-zinc-500 block">Original Language</span>
-                <span className="text-zinc-300">{tmdb.original_language.toUpperCase()}</span>
-              </div>
-            )}
-            {tmdb.production_countries?.length > 0 && (
-              <div>
-                <span className="text-zinc-500 block">Country</span>
-                <span className="text-zinc-300">{tmdb.production_countries.map(c => c.name).join(", ")}</span>
-              </div>
-            )}
-            {tmdb.spoken_languages?.length > 0 && (
-              <div>
-                <span className="text-zinc-500 block">Languages</span>
-                <span className="text-zinc-300">{tmdb.spoken_languages.map(l => l.english_name).join(", ")}</span>
-              </div>
-            )}
-            {tmdb.production_companies?.length > 0 && (
-              <div className="col-span-2 sm:col-span-3">
-                <span className="text-zinc-500 block">Production</span>
-                <span className="text-zinc-300">{tmdb.production_companies.map(c => c.name).join(", ")}</span>
-              </div>
-            )}
-          </div>
-        </Section>
-      )}
-    </div>
-  );
-}
-
-// ─── Utilities ───────────────────────────────────────────────────────────────
-
-function getCertification(results: { iso_3166_1: string; rating: string }[] | undefined, country: string): string | null {
-  if (!results) return null;
-  const match = results.find(r => r.iso_3166_1 === country) || results.find(r => r.iso_3166_1 === "US");
-  return match?.rating || null;
-}
-
-const MONETIZATION_ORDER = [
-  { type: "FLATRATE", label: "Stream" },
-  { type: "FREE", label: "Free" },
-  { type: "ADS", label: "Ads" },
-  { type: "RENT", label: "Rent" },
-  { type: "BUY", label: "Buy" },
-] as const;
-
-function groupOffersByType(offers: Title["offers"]) {
-  const groups: { type: string; label: string; offers: Title["offers"] }[] = [];
-  for (const { type, label } of MONETIZATION_ORDER) {
-    const deduped = new Map<number, Title["offers"][0]>();
-    for (const o of offers) {
-      if (o.monetization_type === type && !deduped.has(o.provider_id)) {
-        deduped.set(o.provider_id, o);
-      }
-    }
-    if (deduped.size > 0) {
-      groups.push({ type, label, offers: Array.from(deduped.values()) });
-    }
-  }
-  return groups;
 }
